% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angle_calculation.R
\name{angle_alpha}
\alias{angle_alpha}
\title{Core angle calculation function}
\usage{
angle_alpha(data, alpha, start)
}
\arguments{
\item{data}{Matrix with 2 columns containing x and y coordinates}

\item{alpha}{Numeric value between 0 and 1 for quantile threshold}

\item{start}{Numeric vector of length 2 for starting point coordinates}
}
\value{
A list containing all angle calculation methods
}
\description{
This is the main function that performs the angle calculations.
It implements the algorithm from the original R files.
}
\keyword{internal}
